+++ cmd/et-core/main.go
@@ -30,6 +30,7 @@ import (
 	_ "github.com/emergency-tunnel/et/internal/transport/stealth"
 	_ "github.com/emergency-tunnel/et/internal/transport/tcp"
 	_ "github.com/emergency-tunnel/et/internal/transport/ws"
+	_ "github.com/emergency-tunnel/et/internal/transport/wss"
 )
 
 func main() {